Revert "Restructuring files"
