+++ package.json
@@ -11,32 +11,31 @@
   "dependencies": {
     "@emotion/react": "^11.13.0",
     "@emotion/styled": "^11.13.0",
-    "@mui/icons-material": "^6.2.0",
-    "@mui/material": "^6.2.0",
-    "@mui/system": "^6.2.0",
+    "@mui/icons-material": "^5.16.5",
+    "@mui/material": "^5.16.5",
+    "@mui/system": "^5.15.15",
     "@next-auth/mongodb-adapter": "^1.1.3",
     "@types/mongodb": "^4.0.7",
     "@types/next-auth": "^3.15.0",
     "mongodb": "^6.7.0",
-    "next": "15.1.0",
+    "next": "14.2.3",
     "next-auth": "^4.24.7",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "^18",
+    "react-dom": "^18",
     "react-icons": "^5.2.1",
-    "react-router-dom": "^7.0.2",
-    "slate": "^0.112.0",
-    "slate-dom": "^0.111.0",
-    "slate-react": "^0.112.0",
+    "react-router-dom": "^6.23.1",
+    "slate": "^0.103.0",
+    "slate-react": "^0.104.0",
     "tsparticles": "^3.5.0",
-    "uuid": "^11.0.3"
+    "uuid": "^9.0.1"
   },
   "devDependencies": {
-    "@types/node": "^22.10.2",
-    "@types/react": "^19.0.1",
-    "@types/react-dom": "^19.0.2",
-    "@types/uuid": "^10.0.0",
-    "eslint": "^9.17.0",
-    "eslint-config-next": "15.1.0",
+    "@types/node": "^20",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/uuid": "^9.0.8",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.3",
     "typescript": "^5"
   }
 }
