+++ src/app/_components/Poem/Poem.tsx
@@ -1,48 +1,20 @@
 import React from "react";
-import { Container, Typography, Box } from "@mui/material";
-import Stanza from "./subcomponents/Stanza/Stanza";
-import { PoemType } from "@/app/_types/Types";
+import { Container } from "@mui/material";
+import Stanza from "./Stanza/Stanza";
+import { PoemStanzaType } from "@/app/_types/Types";
 
 interface PoemProps {
-  poemData: PoemType;
+  stanzas: PoemStanzaType[];
 }
 
-const Poem: React.FC<PoemProps> = ({ poemData }) => {
-  return poemData ? (
-    <>
-      <Typography
-        variant="h4"
-        sx={{
-          wordWrap: "break-word",
-          overflowWrap: "break-word",
-          width: "100%",
-          textAlign: "left",
-          display: "flex",
-          justifyContent: "start",
-          pl: ".2em",
-        }}
-      >
-        {poemData.title}
-      </Typography>
-      <Typography
-        variant="body2"
-        color="textSecondary"
-        sx={{
-          width: "100%",
-          textAlign: "left",
-          pl: ".7em",
-        }}
-      >
-        {poemData.type === "Original" ? poemData.type : null} by{" "}
-        {poemData.author}
-      </Typography>
-      <Container sx={{ py: "1em", pl: ".7em" }} disableGutters>
-        {poemData.stanzas.map((stanza) => (
-          <Stanza key={stanza.id} stanza={stanza} />
-        ))}
-      </Container>
-    </>
-  ) : null;
+const Poem: React.FC<PoemProps> = ({ stanzas }) => {
+  return (
+    <Container maxWidth="sm" sx={{ padding: "10px", paddingLeft: "30px" }}>
+      {stanzas.map((stanza) => (
+        <Stanza key={stanza.id} stanza={stanza} />
+      ))}
+    </Container>
+  );
 };
 
 export default Poem;
