+++ src/app/_utils/typeGuards.ts
@@ -1,0 +1,7 @@
+import { CustomElement, Descendant } from "../_types/Types" 
+
+export function isCustomElement(
+  descendant: Descendant
+): descendant is CustomElement {
+  return "type" in descendant && "children" in descendant;
+}
