+++ src/app/layout.tsx
@@ -1,58 +1,66 @@
 // src/app/layout.tsx
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
+import dynamic from "next/dynamic";
 import "./globals.css";
-import { Box, Container, Grid2 as Grid } from "@mui/material";
-import ClientProviders from "./_components/ClientProviders/ClientProviders";
-import Navbar from "./Navigation/NavBar";
+import { Box, Container, Grid } from "@mui/material";
 
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-	title: "Poem Journal",
-	description: "A journal for poems and text editing",
+  title: "Poem Journal",
+  description: "A journal for poems and text editing",
 };
 
+// Dynamically import ClientProviders with no SSR
+const ClientProviders = dynamic(
+  () => import("@/app/_components/ClientProviders/ClientProviders"),
+  { ssr: false }
+);
+
+// Dynamically import Navbar with no SSR
+const Navbar = dynamic(() => import("./Navigation/NavBar"), { ssr: false });
+
 export default function RootLayout({
-	children,
+  children,
 }: {
-	children: React.ReactNode;
+  children: React.ReactNode;
 }) {
-	return (
-		<html lang="en">
-			<body className={inter.className}>
-				<ClientProviders>
-					<Box
-						sx={{
-							width: "100%",
-							minHeight: "100vh",
-							display: "flex",
-							flexDirection: "column",
-							bgcolor: "secondary",
-						}}
-					>
-						<Navbar />
-						<Container
-							sx={{
-								maxWidth: "xl",
-								flexGrow: 1,
-								display: "flex",
-								justifyContent: "center",
-								alignItems: "center",
-								paddingTop: "60px",
-							}}
-						>
-							<Grid container justifyContent="center" alignItems="center">
-								<Grid xs={12}>
-									<Grid container justifyContent="center" alignItems="unset">
-										{children}
-									</Grid>
-								</Grid>
-							</Grid>
-						</Container>
-					</Box>
-				</ClientProviders>
-			</body>
-		</html>
-	);
+  return (
+    <html lang="en">
+      <body className={inter.className}>
+        <ClientProviders>
+          <Box
+            sx={{
+              width: "100%",
+              minHeight: "100vh",
+              display: "flex",
+              flexDirection: "column",
+              bgcolor: "secondary",
+            }}
+          >
+            <Navbar />
+            <Container
+              sx={{
+                maxWidth: "xl",
+                flexGrow: 1,
+                display: "flex",
+                justifyContent: "center",
+                alignItems: "center",
+                paddingTop: "60px",
+              }}
+            >
+              <Grid container justifyContent="center" alignItems="center">
+                <Grid item xs={12}>
+                  <Grid container justifyContent="center" alignItems="unset">
+                    {children}
+                  </Grid>
+                </Grid>
+              </Grid>
+            </Container>
+          </Box>
+        </ClientProviders>
+      </body>
+    </html>
+  );
 }
