--- src/app/poem/upload/page.tsx
@@ -1,12 +1,0 @@
-import PoemForm from "@/app/_components/Poem/PoemForm/PoemForm";
-import { EditorProvider } from "@/editor/contexts/EditorContext";
-
-function Upload() {
-	return (
-		<EditorProvider>
-			<PoemForm />
-		</EditorProvider>
-	);
-}
-
-export default Upload;
